+++ tests/test_download.py
@@ -1,5 +1,6 @@
 from path import Path
-from fordead.theia_preprocess import maja_download, maja_search, patch_merged_scenes
+from fordead.theia_preprocess import maja_download, maja_search, patch_merged_scenes, categorize_search
+import pandas as pd
 
 
 def test_maja_search():
@@ -10,12 +11,30 @@ def test_maja_search():
     df = maja_search(tile, start_date, end_date)
     assert all(df.version == "4-0")
 
+    # two granules (diff. id, diff. date, diff. version)
     start_date = "2024-09-20"
     end_date = "2024-09-26"
     df = maja_search(tile, start_date, end_date)
     assert df.shape[0] == 2
+    assert df.date.unique().shape[0] == 2
+
+    # two splits of scene
+    start_date = "2024-08-29"
+    end_date = "2024-08-30"
+    df = maja_search(tile, start_date, end_date)
+    assert df.shape[0] == 2
+    assert df.date.unique().shape[0] == 1
+
+    # two granules (same id, same date, diff. version)
+    # the latest is kept
+    tile = "T31UGQ"
+    start_date = "2024-09-18"
+    end_date = "2024-09-19"
+    df = maja_search(tile, start_date, end_date)
+    assert df.shape[0] == 1
 
     # test cloud cover limit
+    tile = "T31TGM"
     start_date = "2024-09-20"
     end_date = "2024-09-26"
     df = maja_search(tile, start_date, end_date, lim_perc_cloud=50)
@@ -27,6 +46,76 @@ def test_maja_search():
     df = maja_search(tile, start_date, end_date)
     assert df.empty
 
+
+def test_categorize_search():
+    temp_local = pd.DataFrame(
+        dict(
+            id = ["test_id"],
+            date = ["2016-01-01"],
+            version = ["1-0"],
+            unzip_file = ["test_file"],
+            merged = [False],
+        )
+    )
+    
+    temp_remote = pd.DataFrame(
+        dict(
+            id = ["test_id"],
+            date = ["2016-01-01"],
+            version = ["1-0"],
+            cloud_cover = ["10"],
+            product = ["test_product"],
+        )
+    )
+
+    ##################
+    # one file on date
+    ##################
+
+    # same date, same id, same version
+    local = temp_local.copy() 
+    remote = temp_remote.copy()
+    df = categorize_search(remote, local)
+    assert df.status.unique().tolist() == ["up_to_date"]
+
+    # same date, same id, different version
+    remote.version=["2-0"]
+    df = categorize_search(remote, local)
+    assert df.status.unique().tolist() == ["upgrade"]
+
+    # local-remote: same date, diff. id, diff. version
+    remote.version=["2-0"]
+    remote.id = ["test_id_2"]
+    df = categorize_search(remote, local)
+    assert df.status.tolist() == ["not-in-search", "download"]
+
+    ###################
+    # two files on date
+    ###################
+    local = pd.concat([temp_local.copy(), temp_local.copy()])
+    local.id=["test_id_1", "test_id_2"]
+    local.merged=[True, True]
+    remote = pd.concat([temp_remote.copy(), temp_remote.copy()])
+    remote.id=["test_id_1", "test_id_2"]
+
+    # local-remote: same date, same id, same version
+    df = categorize_search(remote, local)
+    assert df.status.unique().tolist() == ["up_to_date"]
+
+    # local-remote: same date, same id, different version
+    remote.version=["2-0", "1-0"]
+    df = categorize_search(remote, local)
+    assert df.status.unique().tolist() == ["upgrade"]
+
+    # local-remote: same date, different id, different version
+    remote = pd.concat([temp_remote.copy(), temp_remote.copy()])
+    remote.id=["test_id_1", "test_id_3"]
+    df = categorize_search(remote, local)
+    ref = pd.DataFrame(dict(
+        id=["test_id_1", "test_id_2", "test_id_3"],
+        status = ["download", "not-in-search", "download"]
+        ))
+    assert all(df[["id", "status"]] ==  ref)
 
 def test_download(output_dir):
     zip_dir = (output_dir / "download" / "zip").rmtree_p().makedirs_p()
@@ -64,6 +153,15 @@ def test_download(output_dir):
     # bands=["B2", "B3", "CLMR2", "CLMR1"]
     # cloud_min = 100
     # cloud_max = 100
+    # "20160326-103406-538"
+    
+    # # duplicate with different ID and old version not downloadable anymore
+    # tile = "T31TGM"
+    # start_date = "2016-03-26"
+    # end_date = "2016-03-27"
+    # bands=["B2", "B3", "CLMR2", "CLMR1"]
+    # cloud_min = 100
+    # cloud_max = 100
 
     # # 31TGM 2017-06-19 is "tri"plicate with cloud_cover [3, 4] (2 x v1-4 + v4-0)
     # tile = "T31TGM"
@@ -74,20 +172,21 @@ def test_download(output_dir):
     # cloud_max = 100
 
     # # 31TGM 2018-08-11 is duplicate with cloud_cover [41,52] --> became only one tile with cloud_cover 43
-    # # 31TGK 2020-05-22 is duplicate with cloud_cover [11,30] --> became two tiles with [11, 8]
-    tile = "T31TGK"
-    start_date = "2020-05-22"
-    end_date = "2020-05-23"
+    # # 31TGK 2020-05-22 is duplicate with cloud_cover [11,30] --> became two tiles with [11, 8] --> became one tile
+    # # 31TGM 2024-08-29 is scene split with cc [0, 6] for v3-1
+    tile = "T31TGM"
+    start_date = "2024-08-29"
+    end_date = "2024-08-30"
     bands = ["B2", "B3", "CLMR2", "CLMR1"]
-    cloud_min = 10
-    cloud_max = 20
+    cloud_min = 5
+    cloud_max = 100
 
     # # T31TGK 2023-01-12 is small
     # tile="T31TGK",
     # start_date="2023-01-12",
     # end_date="2023-01-13",
 
-    # should download duplicates
+    # should only download one
     downloaded, unzip_files = maja_download(
         tile=tile,
         start_date=start_date,