+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # v1.11.2
 
 ## Add
-- `maja_download`:
+- `maja_download` & `theia_preprocess` (API & CLI):
   - status `not-in-search` when a local scene is not in the search result (either outdated or filtered by cloud cover)
-  - arg `rm_na` in order to remove local files with status `not-in-search`
+  - arg `rm_na` in order to remove local files not available in search results
   - export `maja_download` results to file `{unzip_dir}/{date}_files_status.tsv`, usefull to analyse them in `dry_run` mode before really applying.
 
 ## Fix