+++ CHANGELOG.md
@@ -5,15 +5,18 @@
   - export `maja_download` results to file `{unzip_dir}/{date}_files_status.tsv`, useful to analyse them in `dry_run` mode before really applying.
   - status `remove` when a local scene is not in the search result (either outdated, filtered out by cloud cover) or older version duplicates with same id
   - arg `rm` in order to delete local scene dirs with status `remove`
+  - add status `failed_download` and `failed_remove` to identify if something went wrong
 
 ## Fix
 - `maja_download` error when trying to remove a scene duplicate already merged (issue #48)
 - mutiple scenes with same tile and datetime on geodes (issue #49)
+- empty local and remote table (issue #51)
 
 ## Change
 - `maja_download`:
   - remove ability to keep splits of scenes with cloud cover over limit --> simplify understanding
   - args `upgrade` default value to `False` in coherence with `theia_preprocess`
+  - move retry to `theia_preprocess` so that every product of every tile is tried before retry
 
 # v1.11.1